% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indupdate}
\alias{indupdate}
\title{Update sampling indicator matrix}
\usage{
indupdate(rejmat, indmat, diffmat, c, k)
}
\arguments{
\item{rejmat}{The rejection matrix}

\item{indmat}{The sample indicator matrix}

\item{diffmat}{The difference matrix}

\item{c}{The constant for checking already-rejected values}

\item{k}{The k in k-FWER}
}
\description{
Update sampling indicator matrix
}
