% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmax}
\alias{kmax}
\title{Returns the kth largest value by sorting in place}
\usage{
kmax(x, k)
}
\arguments{
\item{x}{Numeric vector.}

\item{k}{Specifies the kth largest value.}
}
\description{
Returns the kth largest value by sorting in place
}
