% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankconf-package.R
\docType{package}
\name{rankconf-package}
\alias{rankconf-package}
\title{The 'rankconf' package.}
\description{
A package for rank inference
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
