% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankconf.R
\name{rankconf}
\alias{rankconf}
\title{Construct confidence intervals around ranks}
\usage{
rankconf(
  y,
  sig2,
  type = "FDR",
  alpha = 0.05,
  k = NA,
  best = "min",
  thr = parallel::detectCores() - 1
)
}
\arguments{
\item{y}{estimates}

\item{sig2}{variance of estimates}

\item{type}{What error should be controlled?}

\item{alpha}{The rate at which the error should be controlled. 0.05 by default}

\item{k}{For k-FWER, number of errors to be bounded}

\item{best}{"min" for rank 1 to correspond to lowest estimate.
"max" for rank 1 to correspond to greatest estimate.}

\item{thr}{Number of threads to utilize. Number of cores minus one by default}
}
\description{
Construct confidence intervals around ranks
}
