% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rejupdate}
\alias{rejupdate}
\title{Update the symmetric rejection matrix in place based on what values of
the difference matrix are greater than the provided critical value.}
\usage{
rejupdate(rejmat, diffmat, c)
}
\arguments{
\item{rejmat}{The rejection matrix}

\item{diffmat}{The difference matrix}

\item{c}{A constant}
}
\description{
Update the symmetric rejection matrix in place based on what values of
the difference matrix are greater than the provided critical value.
}
