% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankconf.R
\name{rankconf}
\alias{rankconf}
\title{Construct confidence intervals around ranks.}
\usage{
rankconf(
  y,
  sig2,
  type = "PCER",
  method = "NAIVE",
  alpha = 0.05,
  k = 1,
  best = "min",
  thr = 2
)
}
\arguments{
\item{y}{estimates}

\item{sig2}{variance of estimates}

\item{type}{What error should be controlled?}

\item{method}{What method should be used to control error rates?}

\item{alpha}{The rate at which the error should be controlled. 0.05 by default.}

\item{k}{Number of errors to control, used for familywise error rates. 1 by default.}

\item{best}{"min" for rank 1 to correspond to lowest estimate.
"max" for rank 1 to correspond to greatest estimate.}

\item{thr}{Number of threads to use for parallel processing. 2 by default.}
}
\description{
Construct confidence intervals around ranks.
}
