% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigupdate}
\alias{sigupdate}
\title{Update sampling indicator matrix, in place modifications}
\usage{
sigupdate(rejmat, sigmat, diffmat, c, numind, k)
}
\arguments{
\item{rejmat}{The rejection matrix}

\item{diffmat}{The difference matrix}

\item{c}{The constant for checking already-rejected values}

\item{k}{The k in k-FWER}

\item{indmat}{The sample indicator matrix}
}
\description{
Update sampling indicator matrix, in place modifications
}
